+++ webapp/tests/unit/controller/handlebar-components.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for W-097 Phase 1: Reusable handlebars components
  * @description     Tests for component definition and usage functionality
  * @file            webapp/tests/unit/controller/handlebar-components.test.js
- * @version         1.4.7
- * @release         2026-01-07
+ * @version         1.4.8
+ * @release         2026-01-08
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/